new update, just new update

+++ src/main/java/exp/fluffynuar/truedarkness/procedures/CorruptRune1KoghdaSushchnostKhoditPoBlokuProcedure.java
@@ -7,12 +7,20 @@ import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.core.BlockPos;
 
+import exp.fluffynuar.truedarkness.init.TruedarknessModMobEffects;
+
 public class CorruptRune1KoghdaSushchnostKhoditPoBlokuProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z, Entity entity) {
 		if (entity == null)
 			return;
-		if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
-			_entity.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, 120, 1, false, false));
-		world.destroyBlock(BlockPos.containing(x, y, z), false);
+		if (!(entity instanceof LivingEntity _livEnt0 && _livEnt0.hasEffect(TruedarknessModMobEffects.RUNE_PROTECTION.get()))) {
+			if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+				_entity.addEffect(new MobEffectInstance(MobEffects.CONFUSION, 120, 1, false, false));
+			if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+				_entity.addEffect(new MobEffectInstance(MobEffects.DIG_SLOWDOWN, 120, 1, false, false));
+			if (entity instanceof LivingEntity _entity && !_entity.level().isClientSide())
+				_entity.addEffect(new MobEffectInstance(MobEffects.BLINDNESS, 120, 1, false, false));
+			world.destroyBlock(BlockPos.containing(x, y, z), false);
+		}
 	}
 }
